# optimize getFDCount with sync.Pool