<!--
* New contributors are highly encouraged to read our
  [CONTRIBUTING](/CONTRIBUTING.md) documentation.
* Both Contributor and Reviewer Checklists are available at https://github.com/DataDog/datadog-agent/blob/main/docs/dev/contributing.md#pull-requests.
* The pull request:
  * Should only fix one issue or add one feature at a time.
  * Must update the test suite for the relevant functionality.
  * Should pass all status checks before being reviewed or merged.
* Commit titles should be prefixed with general area of pull request's change.
-->
### What does this PR do?

Optimizes `getFDCount` by using a `sync.Pool` for the buffer passed to `syscall.Readdirent`.

### Motivation

High number of allocations from this code path.

### Additional Notes

I didn't use a single static slice, because I wasn't sure of the concurrency requirements of this code.

BEFORE
```
goos: linux
goarch: arm64
pkg: github.com/DataDog/datadog-agent/pkg/process/procutil BenchmarkGetFDCount
BenchmarkGetFDCount/self_proc
BenchmarkGetFDCount/self_proc-4         	   53605	     22115 ns/op	    8536 B/op	       7 allocs/op
BenchmarkGetFDCount/self_proc-4         	   54898	     22029 ns/op	    8536 B/op	       7 allocs/op
BenchmarkGetFDCount/self_proc-4         	   55078	     22459 ns/op	    8536 B/op	       7 allocs/op
BenchmarkGetFDCount/self_proc-4         	   54662	     21829 ns/op	    8536 B/op	       7 allocs/op
BenchmarkGetFDCount/self_proc-4         	   54435	     22127 ns/op	    8536 B/op	       7 allocs/op
```

AFTER
```
goos: linux
goarch: arm64
pkg: github.com/DataDog/datadog-agent/pkg/process/procutil BenchmarkGetFDCount
BenchmarkGetFDCount/self_proc
BenchmarkGetFDCount/self_proc-4         	   56757	     21069 ns/op	     344 B/op	       6 allocs/op
BenchmarkGetFDCount/self_proc-4         	   56272	     21257 ns/op	     344 B/op	       6 allocs/op
BenchmarkGetFDCount/self_proc-4         	   56191	     21107 ns/op	     344 B/op	       6 allocs/op
BenchmarkGetFDCount/self_proc-4         	   56752	     21188 ns/op	     344 B/op	       6 allocs/op
BenchmarkGetFDCount/self_proc-4         	   57037	     21043 ns/op	     344 B/op	       6 allocs/op
```

### Possible Drawbacks / Trade-offs

<!--
* What are the possible side-effects or negative impacts of the code change?
-->

### Describe how to test/QA your changes

<!--
* Write here in detail or link to detailed instructions on how this change can
  be tested/QAd/validated, including any environment setup.
-->